feature: well, I set up the API and the calls to the required endpoints for now, which are getting all random recipes and getting recipes by specific tags, I made calls to the endpoints but no data was returned to the view, I enabled logging by adding the logging-interceptor lib but it still didn't show in the view, although the logs showed data being returned

+++ app/src/main/java/dev/olaore/recipeze/models/network/NetworkRecipeIngredient.kt
@@ -2,7 +2,7 @@ package dev.olaore.recipeze.models.network
 
 data class NetworkRecipeIngredient (
     var aisle: String,
-    var amount: Int,
+    var amount: Float,
     var consitency: String,
     var id: Int,
     var image: String,
@@ -14,7 +14,7 @@ data class NetworkRecipeIngredient (
 )
 
 data class NetworkRecipeIngredientMeasure(
-    var amount: Int,
+    var amount: Float,
     var unitLong: String,
     var unitShort: String
 )
